+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to NRV are sumed up in this file.
 
 ## [1.1.3] - XXXX-XX-XX
 ### Added
-
+- improved `eval` in `_FEMResults` for serialized calls (added state variables)
 
 ### Fixed
+
 
 
 ### Removed